+++ projects/gleks/ui/src/public-api.ts
@@ -3,6 +3,7 @@
  */
 
 export * from './lib/components/button/button.component';
+export * from './lib/components/button/button.directive';
 export * from './lib/components/accordion/accordion.component';
 export * from './lib/components/collapsible/collapsible.component';
 export * from './lib/components/collapsible/collapsible-trigger.directive';